feat: cleanup + restructuring
