+++ api/src/repositories/telemetry-repositories/telemetry-vendor-repository.interface.ts
